pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "automatic_scientific_qm"
version = "0.1.0"
dependencies = [
    "doc2json @ git+https://github.com/allenai/s2orc-doc2json.git",
    "adapters==1.0.1",
    "aiohappyeyeballs==2.4.4",
    "aiohttp==3.11.11",
    "aiosignal==1.3.2",
    "annotated-types==0.7.0",
    "anthropic==0.45.2",
    "attrs==24.3.0",
    "backoff==2.2.1",
    "beautifulsoup4==4.12.2",
    "black==25.1.0",
    "bs4==0.0.2",
    "certifi==2024.12.14",
    "cfgv==3.4.0",
    "charset-normalizer==3.4.1",
    "click==8.1.8",
    "datasets==3.2.0",
    "dill==0.3.8",
    "distlib==0.3.9",
    "docker-pycreds==0.4.0",
    "filelock==3.17.0",
    "frozenlist==1.5.0",
    "fsspec==2024.9.0",
    "fuzzywuzzy==0.18.0",
    "gdown==5.2.0", 
    "gitdb==4.0.12",
    "GitPython==3.1.44",
    "huggingface-hub==0.27.1",
    "identify==2.6.6",
    "idna==3.10",
    "Jinja2==3.1.5",
    "joblib==1.4.2",
    "Levenshtein==0.26.1",
    "lightning-utilities==0.11.9",
    "lxml==5.3.0",
    "MarkupSafe==3.0.2",
    "mpmath==1.3.0",
    "multidict==6.1.0",
    "multiprocess==0.70.16",
    "mypy==1.14.1",
    "mypy-extensions==1.0.0",
    "networkx==3.4.2",
    "nltk==3.9.1",
    "nodeenv==1.9.1",
    "numpy==2.2.2",
    "nvidia-cublas-cu12==12.4.5.8",
    "nvidia-cuda-cupti-cu12==12.4.127",
    "nvidia-cuda-nvrtc-cu12==12.4.127",
    "nvidia-cuda-runtime-cu12==12.4.127",
    "nvidia-cudnn-cu12==9.1.0.70",
    "nvidia-cufft-cu12==11.2.1.3",
    "nvidia-curand-cu12==10.3.5.147",
    "nvidia-cusolver-cu12==11.6.1.9",
    "nvidia-cusparse-cu12==12.3.1.170",
    "nvidia-nccl-cu12==2.21.5",
    "nvidia-nvjitlink-cu12==12.4.127",
    "nvidia-nvtx-cu12==12.4.127",
    "openai==1.61.0",
    "packaging==24.2",
    "pandas==2.2.3",
    "pathspec==0.12.1",
    "platformdirs==4.3.6",
    "pre_commit==4.1.0",
    "propcache==0.2.1",
    "protobuf==5.29.3",
    "psutil==6.1.1",
    "pyarrow==19.0.0",
    "pydantic==2.10.5",
    "pydantic_core==2.27.2",
    "python-dateutil==2.9.0.post0",
    "python-Levenshtein==0.26.1",
    "pytorch-lightning==2.5.0.post0",
    "pytz==2024.2",
    "PyYAML==6.0.2",
    "RapidFuzz==3.12.1",
    "regex==2024.11.6",
    "requests==2.32.3",
    "retry==0.9.2",
    "safetensors==0.5.2",
    "sentry-sdk==2.20.0",
    "setproctitle==1.3.4",
    "six==1.17.0",
    "smmap==5.0.2",
    "soupsieve==2.6",
    "sympy==1.13.1",
    "tokenizers==0.20.3",
    "torch==2.5.1",
    "torchmetrics==1.6.1",
    "torchtyping==0.1.5",
    "tqdm==4.67.1",
    "transformers==4.45.2",
    "triton==3.1.0",
    "typeguard==2.13.3",
    "typing_extensions==4.12.2",
    "tzdata==2025.1",
    "urllib3==2.3.0",
    "virtualenv==20.29.1",
    "wandb==0.19.4",
    "xxhash==3.5.0",
    "yarl==1.18.3"
]
requires-python = ">=3.11"
[project.optional-dependencies]
dev = [
    "black"
]
